add twc_core & twc_curator
moved some functions over to core already, curator will need a decent
clean & prep + readme.mds explaining each pbo
