+++ @twc_framework/addons/twc_core/CfgFunctions.hpp
@@ -1,0 +1,21 @@
+class CfgFunctions {
+	class TWC {
+		class functions {
+			file = "twc_core\functions";
+			
+			/** MISC FUNCTIONALITY **/
+			class soundLoop {};
+			
+			/** PLAYER INFORMATION CHECKS **/
+			class isMember {};
+			class isManagement {};
+			class isHigherRank {};
+			//class isHighestRank {};
+			//class isInTopRanks {}; // variant of above, to account for any potential 2ic
+			
+			/** UTILITY FUNCTIONS **/
+			class findGreatestDistanceWithArray {};
+			class findGreatestDistancePairInArray {};
+		};
+	};
+};
