+++ @twc_framework/addons/twc_core/README.md
@@ -1,0 +1,4 @@
+# TWC_Core
+TWC Core offers utility functions that can be utilised anywhere. Generic functionality that might be commonplace. For example, the function "isMember", or alternatively "isHighestRank".
+
+The idea is to keep these out of the framework, to remove bloat from there, and to serve as an easy place to at glance to find functions to help you out writing other ones. Less repeating code.
