+++ @twc_framework/addons/twc_core/config.cpp
@@ -1,0 +1,18 @@
+class CfgPatches {
+	class TWC_Core {
+		units[] = {};
+		weapons[] = {};
+		requiredVersion = 1.7;
+		
+		requiredAddons[] = {"cba_xeh"};
+		
+		author[] = {"Bosenator"};
+		authorUrl = "thewreckingcrew.eu";
+		version = "1";
+		versionStr = "1";
+		versionAr[] = {1};
+	};
+};
+
+#include "CfgEventHandlers.hpp"
+#include "CfgFunctions.hpp"
