+++ @twc_framework/addons/twc_core/functions/fn_findGreatestDistanceWithArray.sqf
