+++ @twc_framework/addons/twc_core/functions/fn_isMember.sqf
