+++ @twc_framework/addons/twc_curator/README.md
@@ -1,0 +1,2 @@
+# TWC_Curator
+All Curator/Zeus functionality is located here. Framework had become bloated, and so these were relocated here for ease.
